+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Full documentation for hipCUB is available at [https://rocm.docs.amd.com/project
 * Added an overload of `WarpScan::InclusiveScan` that accepts an initial value to seed the scan.
 * `UnrolledThreadLoad`, `UnrolledCopy`, and `ThreadLoadVolatilePointer` were added to align hipCUB with CUB.
 * `ThreadStoreVolatilePtr` and the `IterateThreadStore` struct were added to align hipCUB with CUB.
+* Added `hipcub::InclusiveScanInit` for CUB parity.
 
 ### Removed